+++ package.json
@@ -21,6 +21,7 @@
     "date-fns": "^2.30.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "gsap": "^3.11.5",
     "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
