+++ src/components/utils/types.ts
@@ -1,0 +1,4 @@
+export type ChartData = {
+  x: string,
+  y: number,
+}
